feat: move config and questions to their own crates
